pyproject.toml
[project]
name = "healthclaw-guardrails"
version = "1.3.0"
description = "HealthClaw Guardrails — FHIR R4 US Core v9 + R6 MCP guardrail proxy (healthclaw.io)"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.27.0",
    "marshmallow>=3.26.1",
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0",
    "redis>=4.6.0",
    "requests>=2.32.3",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
